Improve the volume smoothing algorithm.
It is now much more uniform across devices where the maximum volume
differs, as it operates on percentages instead of volume counts.

The update rate has been increased (and approach rate adjusted) for
smoother adjustments. To make sure we're not needlessly spinning the
CPU, the thread will sleep once it has reached its target, and will
be awoken when a new target volume is set.