Translations suggestions for english
Edits mentionned in #28.
Added a few more, mostly to clarify and homogenize maximum and minimum volume adjustments sections.
I removed the word Optional from one of the latest strings. I felt it was the more complicated option in term of formulation, and that it was enough obvious this option is optional.